+++ apps/cookbook/src/app/examples/stock-chart-example/example/comparison.ts
