+++ apps/cookbook/src/app/examples/stock-chart-example/stock-chart-example.component.html
@@ -1,5 +1,5 @@
 <kirby-card [hasPadding]="true">
-  <cookbook-stock-chart-example-simple></cookbook-stock-chart-example-simple>
+  <cookbook-stock-chart-example-simple-dataset></cookbook-stock-chart-example-simple-dataset>
 </kirby-card>
 
 <kirby-card [hasPadding]="true">
