+++ apps/cookbook/src/app/examples/stock-chart-example/stock-chart-example.component.ts
@@ -1,12 +1,16 @@
 import { Component } from '@angular/core';
 import { CardModule } from '@kirbydesign/designsystem/card';
-import { StockChartExampleComparisonComponent } from './example/stock-chart-example-comparison.component';
-import { StockChartExampleSimpleComponent } from './example/stock-chart-example-simple.component';
+import { StockChartExampleComparisonComponent } from './example/comparison';
+import { StockChartExampleSimpleDatasetComponent } from './example/simple-dataset';
 
 @Component({
   selector: 'cookbook-stock-chart-example',
   templateUrl: './stock-chart-example.component.html',
   styleUrls: ['./stock-chart-example.component.scss'],
-  imports: [CardModule, StockChartExampleComparisonComponent, StockChartExampleSimpleComponent],
+  imports: [
+    CardModule,
+    StockChartExampleComparisonComponent,
+    StockChartExampleSimpleDatasetComponent,
+  ],
 })
 export class StockChartExampleComponent {}
